fix(api): add run_server to app.py
